requirements.txt
annotated-types==0.7.0
anyio==4.4.0
blinker==1.8.2
certifi==2024.7.4
click==8.1.7
exceptiongroup==1.2.2
flask[async]==3.0.3
h11==0.14.0
httpcore==1.0.5
httpx==0.25.2
httpx-sse==0.3.1
idna==3.7
importlib_metadata==8.2.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
pip==21.2.4
pydantic==2.8.2
pydantic_core==2.20.1
python-dotenv==1.0.1
setuptools==58.0.4
sniffio==1.3.1
typing_extensions==4.12.2
Werkzeug==3.0.3
zipp==3.19.2
requests==2.32.3
aiohttp==3.9.5